Scheme Model
Stretching the Scheme class as much as is reasonable. The class will handle the local logic of schemes; details that will be identical from scheme to scheme, for the most part. The scheming service will handle the finer details of manipulating and examining schemes. This disctinction should clean up the code and provide some ease to expansion. Next up is incorporating the cash cost into Beginner Lair (and then perhaps other places)