Recruit Model
Started building the recruit model with the thought that it could be constructed by the player service. This is problematic because it needs to reference the player service. Consider reducing the imports based on model from the whole service to just the relevant parts. Snapshot because I gotta go afk. (Recruit model is currently waay wrong)