Recruitment Model
Made the recruitment model, which is cool. Excited because I discovered that this setup means that the components do not need to reference the player service. Going to move forward and attempt to continue this trend, and make it so the core services are only referenced by lesser services and not templates.