Some work on Training
It's really bugging me that I have to include the entire numbers and player services in the models. To the point that I'm going to rethink my approach. I took a peek into the memory usage of the site in its current state and though it appears fine, I have no idea how it will scale. It's just downright sloppy, so I'll probably pivot to working on the save functionality since that's the driver of this complication.