Base Class
Here we have a Base Class that is going to do the legwork of the player service. Having these variables available everywhere was always the goal, so this is the approach that currently makes most sense.

I'm about ready to start obliterating a lot of old stuff with a switch over to important details being held inside the scheme array instead of elsewhere, so things will be broken for a while. This commit is stable and just does some extra behind the scenes work that isn't used by anything yet.