Base Service - Schemes mostly complete
I should have done this ages ago. By having the core gameplay items on a Base class, as well as utilizing it for constructing the templates and models, the data that was previously available on the player service is now readily available to all components and services. This simplifies a lot of previous pain points, and enables rational model construction and usage. Schemes themselves belong at the very base class alongside other passive modifiers. The second layer will perform calculations and identify the resulting numbers (in a way that is available for diagnostic output for both myself and players). This whole structure feels a lot better.

This commit is running smoothly with schemes and scheme-related objects residing on the base class. The game auto-saves every minute, and will remember scheme progress as well as current scheme (opening the preview pane if one is selected) if you were actively earning scheme points when you saved.