Recruits appears to be working
This is a cautious commit, but the save data is accurately retaining simple recruitment object data. Alas, I had to shed a lot of the convenience in the model thanks to the new setup, but with recruiting service handling most things via methods with passed in recruit objects, it really amounts to the same thing, even if the syntax is in a less preferred form. Going to tinker some more, but I believe this is solid.

Includes a few balance tweaks to recruitment objects and lodging. I want henchmen to accrue slower than they do now, and I want the lodging scheme to completely remove the worry of capping out on henchmen, at least in the early game. Feels a lot better with the recruitment objects spending more time progressing visibly on the screen.