Readout Modal for Base details
Right now it is displaying dummy data, but I'm liking how it looks and the necessary pieces are available for Guards to actually have an impact on the previously non-existant Lair stats. Once the save/load functionality is fully coded, sorting things out so they are in a logical order will be very helpful.