Guard Duty under NNWO and a Lair Modal
So Guards didn't actually do anything, just had a couple of variables that remembered things about them. While applying the changes to Guards, I decided to start in on the Lair itself and some of its stats, since Guards will impact that. Right now it's some arbitrary bindings and references, but it was used with the Lair Modal to give a readout to better describe what Guards are actually doing for the player. I tinkered with removing the Guard inventory readout in the resources area, but I think all advanced henchmen should show up in there, so I'm leaving it in for now.

Not thoroughly tested, but appears to be working well. This massive load logic will need to be swiftly corrected, perhaps before operations.