The Demise of the Numbers Service
With preliminary work on the operations service, the Numbers Service is gone. All relevant items from it have migrated to better locations (mostly BaseNum, a few of the building block pieces made it to Base) and so components that utilized this service were updated to no longer construct with it.

Operations are coming along fine, though they are a little different than training or recruitment objects. As of now the site is configured to delete your save and create a new session each time as Operations are only supported by the 'no save data found' loop. I have some IRL notes on how to change around heists and shady-business-deals to make them feel unique and provide different gameplay options. This will be worked on after my interview this afternoon. Woohoo!