Operations NNWO
Working on Operations and also styling a new way to preview operations. Their new structure requires this, and it was getting a bit bunchy.

The requires column will show icons of the resources that the operation costs. Sliders will be present at the bottom, and once any amount of resources are allocated, an option will appear to operate.

Risk works the same but is moved. Unsure if I want to change it just yet.

Rewards now outputs the 'formula' for how much of whatever it is they'll get. I settled on using 'd6, d12' etc. notation rather than an FA icon since no good ones were avialable. Probably going to need some ngSwitch for this section.

Success (like rewards) is just a raw output of what it might look like when the chance of success is 0%. This will need to be bound to the sliders and a bunch of styling logic.

Also have been building out heists only. A lot of the bulider functions will need to be expanded so they can work for multiple types of operations.