Slider for Operations
After much struggling with Angular Material, got a basic slider attached to the operation and rigged up to the template. A new scheme button was created, but the logic for actually scheming is untouched. This is mostly a baseline proof-of-concept implementation, but lots of relevant items captured in the parking lot.