Operations
Lots of changes to the display and function. Heists are essentially ready to go with everything in place. I am now extending a lot of the operating methods to be able to be used for different types of operations (following the template as a guide for what is next). I enabled the ability to indicate multiple resources as a cost (money and henchmen for shady-business-deals) so I'm going to break and figure out how that should work. Need to come up with what money is used for beyond schemes.