Operations finalized
Sorted out a lot of nonsense remaining in the operating service. (Don't worry, there is plenty of nonsense left - it's just working nonsense). Operations do what they're supposed to and now it's a matter of balancing. On-screen elements are still not very well planned, but usability is king for now.

Shady Business Deals are neat, but need to be supported with something that reports passive income amounts. Passive income is also a save item which needs to be included. On that note, none of this new functionality is actually supported by saving as it wipes the save on open. That'll need immediate attention before operations can be completely closed out.