Baseline Notoriety Meter Support
Rewired the Notoriety Meter like it was before, with a bit more precision and style. Floating point numbers should no longer be a concern with a function on Base itself.

Other than that we've got some day's-work stuff going on here. Added support for the final schemes in the lair tree, though both are not complete just yet. Moved around some of the names of stats within the Lair area to be a bit more flavorful.

Rapidly approaching heroes. . .