New Scheme: Rocket Science
Rocket Science enables players to bank scheme points earned while not scheming. As it levels, a higher percentage of these points are banked (this means some functions to handle rounding). As with other Schemes at this tier that will bank scheme points, it will also have an impact on how many are distributed each second. As of right now each level banks 2% of 'lost' scheme points and cashes out 1 banked point per second. Easily tweakable if needded.

I'm also changing how I store mongoDB data. It's moved to the mongo-data folder to be more obvious in the repo. This means that the accompanying mongodb data may be committed alongside the relevant code, like I'm doing here.

Beyond that I removed some console logs and added a readout for banked points. It'll go away later (moved to the scheming modal) but it's handy right now for debugging.